+++ tests/MyTestSuiteWeb.yml
@@ -1,6 +1,6 @@
 name: TestSuite Web
 testcases:
-- name: TestCase Get URL and check totle
+- name: TestCase Get URL and check title
   context:
     type: web
     width: 1920
